Time: 1 ms (98.25%), Space: 43.3 MB (54.44%) - LeetHub

+++ 3110-score-of-a-string/3110-score-of-a-string.java
@@ -2,7 +2,7 @@ class Solution {
     public int scoreOfString(String s) {
         int c=0;
         for(int i=0;i<s.length()-1;i++){
-            c+=Math.abs(s.charAt(i)-s.charAt(i+1));
+            c+=Math.abs((int)s.charAt(i)-(int)s.charAt(i+1));
         }
         return c;
     }
